requirements.txt
# Core AI framework
crewai[tools]

# LLM & RAG
langchain-community
langchain-openai
langchain-text-splitters
chromadb

# HTTP client for API services
httpx

# Backend API
fastapi
uvicorn
websockets

# Utilities
python-dotenv
pydantic
